refactor: 컴포넌트 폴더구조 개선작업
